+++ hutool-core/src/main/java/com/xiaoleilu/hutool/codec/BCD.java
@@ -1,4 +1,4 @@
-package com.xiaoleilu.hutool.lang;
+package com.xiaoleilu.hutool.codec;
 
 /**
  * BCD码（Binary-Coded Decimal‎）亦称二进码十进数或二-十进制代码<br>