+++ hutool-core/src/main/java/com/xiaoleilu/hutool/util/EnumUtil.java
@@ -4,6 +4,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
+import com.xiaoleilu.hutool.map.MapUtil;
+
 /**
  * 枚举工具类
  *