+++ hutool-core/src/test/java/com/xiaoleilu/hutool/collection/CollUtilTest.java
@@ -14,8 +14,8 @@ import com.xiaoleilu.hutool.collection.CollUtil.Hash;
 import com.xiaoleilu.hutool.lang.Dict;
 import com.xiaoleilu.hutool.lang.Editor;
 import com.xiaoleilu.hutool.lang.Matcher;
+import com.xiaoleilu.hutool.map.MapUtil;
 import com.xiaoleilu.hutool.util.CollectionUtil;
-import com.xiaoleilu.hutool.util.MapUtil;
 
 /**
  * 集合工具类单元测试